refactor(core): killed CSG & CAG : now shape3 & shape2 * split out any remaining functions/methods * most of these are not working yet ! * cleaned up a few helpers * moved debugHelpers to moveOutToCore
