--- src/core/CAGFactories.js
@@ -1,152 +1,0 @@
-const Side = require('./math/Side')
-const Vector2 = require('./math/Vector2')
-const Vertex2 = require('./math/Vertex2')
-const {areaEPS} = require('./constants')
-const {isSelfIntersecting} = require('./utils/cagValidation')
-const canonicalize = require('./utils/canonicalize')
-const {area} = require('./utils/cagMeasurements')
-
-/** Construct a CAG from a list of `Side` instances.
- * this is a duplicate of CAG's fromSides to avoid circular dependency CAG => fromSides => CAG
- * @param {Side[]} sides - list of sides
- * @returns {CAG} new CAG object
- */
-const fromSides = function (sides) {
-  const CAG = require('./CAG')
-  let cag = new CAG()
-  cag.sides = sides
-  return cag
-}
-
-// Converts a CSG to a  The CSG must consist of polygons with only z coordinates +1 and -1
-// as constructed by _toCSGWall(-1, 1). This is so we can use the 3D union(), intersect() etc
-const fromFakeCSG = function (csg) {
-  let sides = csg.polygons.map(function (p) {
-    return Side._fromFakePolygon(p)
-  })
-  .filter(function (s) {
-    return s !== null
-  })
-  return fromSides(sides)
-}
-
-/** Construct a CAG from a list of points (a polygon).
- * The rotation direction of the points is not relevant.
- * The points can define a convex or a concave polygon.
- * The polygon must not self intersect.
- * @param {points[]} points - list of points in 2D space
- * @returns {CAG} new CAG object
- */
-const fromPoints = function (points) {
-  let numpoints = points.length
-  if (numpoints < 3) throw new Error('CAG shape needs at least 3 points')
-  let sides = []
-  let prevpoint = new Vector2(points[numpoints - 1])
-  let prevvertex = new Vertex2(prevpoint)
-  points.map(function (p) {
-    let point = new Vector2(p)
-    let vertex = new Vertex2(point)
-    let side = new Side(prevvertex, vertex)
-    sides.push(side)
-    prevvertex = vertex
-  })
-  let result = fromSides(sides)
-  if (isSelfIntersecting(result)) {
-    throw new Error('Polygon is self intersecting!')
-  }
-  let resultArea = area(result)
-  if (Math.abs(resultArea) < areaEPS) {
-    throw new Error('Degenerate polygon!')
-  }
-  if (resultArea < 0) {
-    result = result.flipped()
-  }
-  result = canonicalize(result)
-  return result
-}
-
-/** Reconstruct a CAG from an object with identical property names.
- * @param {Object} obj - anonymous object, typically from JSON
- * @returns {CAG} new CAG object
- */
-const fromObject = function (obj) {
-  let sides = obj.sides.map(function (s) {
-    return Side.fromObject(s)
-  })
-  let cag = fromSides(sides)
-  cag.isCanonicalized = obj.isCanonicalized
-  return cag
-}
-
-/** Construct a CAG from a list of points (a polygon).
- * Like fromPoints() but does not check if the result is a valid polygon.
- * The points MUST rotate counter clockwise.
- * The points can define a convex or a concave polygon.
- * The polygon must not self intersect.
- * @param {points[]} points - list of points in 2D space
- * @returns {CAG} new CAG object
- */
-const fromPointsNoCheck = function (points) {
-  let sides = []
-  let prevpoint = new Vector2(points[points.length - 1])
-  let prevvertex = new Vertex2(prevpoint)
-  points.map(function (p) {
-    let point = new Vector2(p)
-    let vertex = new Vertex2(point)
-    let side = new Side(prevvertex, vertex)
-    sides.push(side)
-    prevvertex = vertex
-  })
-  return fromSides(sides)
-}
-
-/** Construct a CAG from a 2d-path (a closed sequence of points).
- * Like fromPoints() but does not check if the result is a valid polygon.
- * @param {path} Path2 - a Path2 path
- * @returns {CAG} new CAG object
- */
-const fromPath2 = function (path) {
-  if (!path.isClosed()) throw new Error('The path should be closed!')
-  return fromPoints(path.getPoints())
-}
-
-/** Reconstruct a CAG from the output of toCompactBinary().
- * @param {CompactBinary} bin - see toCompactBinary()
- * @returns {CAG} new CAG object
- */
-const fromCompactBinary = function (bin) {
-  if (bin['class'] !== 'CAG') throw new Error('Not a CAG')
-  let vertices = []
-  let vertexData = bin.vertexData
-  let numvertices = vertexData.length / 2
-  let arrayindex = 0
-  for (let vertexindex = 0; vertexindex < numvertices; vertexindex++) {
-    let x = vertexData[arrayindex++]
-    let y = vertexData[arrayindex++]
-    let pos = new Vector2(x, y)
-    let vertex = new Vertex2(pos)
-    vertices.push(vertex)
-  }
-  let sides = []
-  let numsides = bin.sideVertexIndices.length / 2
-  arrayindex = 0
-  for (let sideindex = 0; sideindex < numsides; sideindex++) {
-    let vertexindex0 = bin.sideVertexIndices[arrayindex++]
-    let vertexindex1 = bin.sideVertexIndices[arrayindex++]
-    let side = new Side(vertices[vertexindex0], vertices[vertexindex1])
-    sides.push(side)
-  }
-  let cag = fromSides(sides)
-  cag.isCanonicalized = true
-  return cag
-}
-
-module.exports = {
-  fromSides,
-  fromObject,
-  fromPoints,
-  fromPointsNoCheck,
-  fromPath2,
-  fromFakeCSG,
-  fromCompactBinary
-}
