--- src/core/CSGFactories.js
@@ -1,121 +1,0 @@
-const Vector3D = require('./math/Vector3')
-const Vertex = require('./math/Vertex3')
-const Plane = require('./math/Plane')
-const Polygon2 = require('./math/Polygon2')
-const Polygon3 = require('./math/Polygon3')
-
-/** Construct a CSG solid from a list of `Polygon` instances.
- * @param {Polygon[]} polygons - list of polygons
- * @returns {CSG} new CSG object
- */
-const fromPolygons = function (polygons) {
-  const CSG = require('./CSG')
-  let csg = new CSG()
-  csg.polygons = polygons
-  csg.isCanonicalized = false
-  csg.isRetesselated = false
-  return csg
-}
-
-/** Construct a CSG solid from a list of pre-generated slices.
- * See Polygon.prototype.solidFromSlices() for details.
- * @param {Object} options - options passed to solidFromSlices()
- * @returns {CSG} new CSG object
- */
-function fromSlices (options) {
-  return Polygon2.createFromPoints([
-        [0, 0, 0],
-        [1, 0, 0],
-        [1, 1, 0],
-        [0, 1, 0]
-  ]).solidFromSlices(options)
-}
-
-/** Reconstruct a CSG solid from an object with identical property names.
- * @param {Object} obj - anonymous object, typically from JSON
- * @returns {CSG} new CSG object
- */
-function fromObject (obj) {
-  let polygons = obj.polygons.map(function (p) {
-    return Polygon3.fromObject(p)
-  })
-  let csg = fromPolygons(polygons)
-  csg.isCanonicalized = obj.isCanonicalized
-  csg.isRetesselated = obj.isRetesselated
-  return csg
-}
-
-/** Reconstruct a CSG from the output of toCompactBinary().
- * @param {CompactBinary} bin - see toCompactBinary().
- * @returns {CSG} new CSG object
- */
-function fromCompactBinary (bin) {
-  if (bin['class'] !== 'CSG') throw new Error('Not a CSG')
-  let planes = []
-  let planeData = bin.planeData
-  let numplanes = planeData.length / 4
-  let arrayindex = 0
-  let x, y, z, w, normal, plane
-  for (let planeindex = 0; planeindex < numplanes; planeindex++) {
-    x = planeData[arrayindex++]
-    y = planeData[arrayindex++]
-    z = planeData[arrayindex++]
-    w = planeData[arrayindex++]
-    normal = Vector3D.Create(x, y, z)
-    plane = new Plane(normal, w)
-    planes.push(plane)
-  }
-
-  let vertices = []
-  const vertexData = bin.vertexData
-  const numvertices = vertexData.length / 3
-  let pos
-  let vertex
-  arrayindex = 0
-  for (let vertexindex = 0; vertexindex < numvertices; vertexindex++) {
-    x = vertexData[arrayindex++]
-    y = vertexData[arrayindex++]
-    z = vertexData[arrayindex++]
-    pos = Vector3D.Create(x, y, z)
-    vertex = new Vertex(pos)
-    vertices.push(vertex)
-  }
-
-  let shareds = bin.shared.map(function (shared) {
-    return Polygon3.Shared.fromObject(shared)
-  })
-
-  let polygons = []
-  let numpolygons = bin.numPolygons
-  let numVerticesPerPolygon = bin.numVerticesPerPolygon
-  let polygonVertices = bin.polygonVertices
-  let polygonPlaneIndexes = bin.polygonPlaneIndexes
-  let polygonSharedIndexes = bin.polygonSharedIndexes
-  let numpolygonvertices
-  let polygonvertices
-  let shared
-  let polygon // already defined plane,
-  arrayindex = 0
-  for (let polygonindex = 0; polygonindex < numpolygons; polygonindex++) {
-    numpolygonvertices = numVerticesPerPolygon[polygonindex]
-    polygonvertices = []
-    for (let i = 0; i < numpolygonvertices; i++) {
-      polygonvertices.push(vertices[polygonVertices[arrayindex++]])
-    }
-    plane = planes[polygonPlaneIndexes[polygonindex]]
-    shared = shareds[polygonSharedIndexes[polygonindex]]
-    polygon = new Polygon3(polygonvertices, shared, plane)
-    polygons.push(polygon)
-  }
-  let csg = fromPolygons(polygons)
-  csg.isCanonicalized = true
-  csg.isRetesselated = true
-  return csg
-}
-
-module.exports = {
-  fromPolygons,
-  fromSlices,
-  fromObject,
-  fromCompactBinary
-}
