--- src/core/CSGToOther.js
@@ -1,138 +1,0 @@
-const canonicalize = require('./utils/canonicalize')
-const Polygon3 = require('./math/Polygon3')
-
-/** returns a compact binary representation of this csg
-   * usually used to transfer CSG objects to/from webworkes
-   * NOTE: very interesting compact format, with a lot of reusable ideas
-   * @returns {Object} compact binary representation of a CSG
-   */
-const toCompactBinary = (_csg) => {
-  let csg = canonicalize(_csg)
-  let numpolygons = csg.polygons.length
-  let numpolygonvertices = 0
-
-  let numvertices = 0
-  let vertexmap = {}
-  let vertices = []
-
-  let numplanes = 0
-  let planemap = {}
-  let planes = []
-
-  let shareds = []
-  let sharedmap = {}
-  let numshared = 0
-        // for (let i = 0, iMax = csg.polygons.length; i < iMax; i++) {
-        //  let p = csg.polygons[i];
-        //  for (let j = 0, jMax = p.length; j < jMax; j++) {
-        //      ++numpolygonvertices;
-        //      let vertextag = p[j].getTag();
-        //      if(!(vertextag in vertexmap)) {
-        //          vertexmap[vertextag] = numvertices++;
-        //          vertices.push(p[j]);
-        //      }
-        //  }
-  csg.polygons.map(function (polygon) {
-      // FIXME: why use map if we do not return anything ?
-      // either for... or forEach
-    polygon.vertices.map(function (vertex) {
-      ++numpolygonvertices
-      let vertextag = vertex.getTag()
-      if (!(vertextag in vertexmap)) {
-        vertexmap[vertextag] = numvertices++
-        vertices.push(vertex)
-      }
-    })
-
-    let planetag = polygon.plane.getTag()
-    if (!(planetag in planemap)) {
-      planemap[planetag] = numplanes++
-      planes.push(polygon.plane)
-    }
-    let sharedtag = polygon.shared.getTag()
-    if (!(sharedtag in sharedmap)) {
-      sharedmap[sharedtag] = numshared++
-      shareds.push(polygon.shared)
-    }
-  })
-
-  let numVerticesPerPolygon = new Uint32Array(numpolygons)
-  let polygonSharedIndexes = new Uint32Array(numpolygons)
-  let polygonVertices = new Uint32Array(numpolygonvertices)
-  let polygonPlaneIndexes = new Uint32Array(numpolygons)
-  let vertexData = new Float64Array(numvertices * 3)
-  let planeData = new Float64Array(numplanes * 4)
-  let polygonVerticesIndex = 0
-
-    // FIXME: doublecheck : why does it go through the whole polygons again?
-    // can we optimise that ? (perhap due to needing size to init buffers above)
-  for (let polygonindex = 0; polygonindex < numpolygons; ++polygonindex) {
-    let polygon = csg.polygons[polygonindex]
-    numVerticesPerPolygon[polygonindex] = polygon.vertices.length
-    polygon.vertices.map(function (vertex) {
-      let vertextag = vertex.getTag()
-      let vertexindex = vertexmap[vertextag]
-      polygonVertices[polygonVerticesIndex++] = vertexindex
-    })
-    let planetag = polygon.plane.getTag()
-    let planeindex = planemap[planetag]
-    polygonPlaneIndexes[polygonindex] = planeindex
-    let sharedtag = polygon.shared.getTag()
-    let sharedindex = sharedmap[sharedtag]
-    polygonSharedIndexes[polygonindex] = sharedindex
-  }
-  let verticesArrayIndex = 0
-  vertices.map(function (vertex) {
-    const pos = vertex.pos
-    vertexData[verticesArrayIndex++] = pos._x
-    vertexData[verticesArrayIndex++] = pos._y
-    vertexData[verticesArrayIndex++] = pos._z
-  })
-  let planesArrayIndex = 0
-  planes.map(function (plane) {
-    const normal = plane.normal
-    planeData[planesArrayIndex++] = normal._x
-    planeData[planesArrayIndex++] = normal._y
-    planeData[planesArrayIndex++] = normal._z
-    planeData[planesArrayIndex++] = plane.w
-  })
-
-  let result = {
-    'class': 'CSG',
-    numPolygons: numpolygons,
-    numVerticesPerPolygon: numVerticesPerPolygon,
-    polygonPlaneIndexes: polygonPlaneIndexes,
-    polygonSharedIndexes: polygonSharedIndexes,
-    polygonVertices: polygonVertices,
-    vertexData: vertexData,
-    planeData: planeData,
-    shared: shareds
-  }
-  return result
-}
-
-/** returns the triangles of this csg
-   * @returns {Polygons} triangulated polygons
-   */
-const toTriangles = (_csg) => {
-  let polygons = []
-  _csg.polygons.forEach(function (poly) {
-    let firstVertex = poly.vertices[0]
-    for (let i = poly.vertices.length - 3; i >= 0; i--) {
-      polygons.push(new Polygon3(
-        [
-          firstVertex,
-          poly.vertices[i + 1],
-          poly.vertices[i + 2]
-        ],
-          poly.shared,
-          poly.plane))
-    }
-  })
-  return polygons
-}
-
-  /** @return {Polygon3[]} The list of polygons. */
-const toPolygons = _csg => _csg.polygons
-
-module.exports = {toCompactBinary, toTriangles, toPolygons}
