+++ src/core/geometry/shape2/fromPointsNoCheck.js
@@ -1,0 +1,21 @@
+/** Construct a CAG from a list of points (a polygon).
+ * Like fromPoints() but does not check if the result is a valid polygon.
+ * The points MUST rotate counter clockwise.
+ * The points can define a convex or a concave polygon.
+ * The polygon must not self intersect.
+ * @param {points[]} points - list of points in 2D space
+ * @returns {CAG} new CAG object
+ */
+const fromPointsNoCheck = function (points) {
+  let sides = []
+  let prevpoint = new Vector2(points[points.length - 1])
+  let prevvertex = new Vertex2(prevpoint)
+  points.map(function (p) {
+    let point = new Vector2(p)
+    let vertex = new Vertex2(point)
+    let side = new Side(prevvertex, vertex)
+    sides.push(side)
+    prevvertex = vertex
+  })
+  return fromSides(sides)
+}
