+++ src/core/geometry/shape2/toCSGWall.js
@@ -1,3 +1,5 @@
+const fromPolygons = require('../shape3/fromPolygons')
+
 /** convert a CAG to a CSG 'wall' of zero thickness
  * @param  {} z0
  * @param  {} z1
@@ -8,3 +10,5 @@ const toCSGWall = function (cag, z0, z1) {
   })
   return fromPolygons(polygons)
 }
+
+module.exports = toCSGWall
