+++ src/core/geometry/shape2/toPoints.js
@@ -16,3 +16,5 @@
     }
     return points
   }
+
+  module.exports = toPoints
