+++ src/core/geometry/shape3/fromPolygons.js
@@ -1,14 +1,15 @@
+const create = require('./create')
+
 /** Construct a CSG solid from a list of `Polygon` instances.
  * @param {Polygon[]} polygons - list of polygons
  * @returns {CSG} new CSG object
  */
 const fromPolygons = function (polygons) {
-  const CSG = require('./CSG')
-  let csg = new CSG()
-  csg.polygons = polygons
-  csg.isCanonicalized = false
-  csg.isRetesselated = false
-  return csg
+  let shape3 = create()
+  shape3.polygons = polygons
+  shape3.isCanonicalized = false
+  shape3.isRetesselated = false
+  return shape3
 }
 
 module.exports = fromPolygons
